mahakim_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pycryptodome==3.19.0
